+++ db/events.php
@@ -36,5 +36,8 @@ $observers = array(
         'eventname'   => '\core\event\course_content_deleted',
         'callback'    => 'format_grid_observer::course_content_deleted',
     ),
-
+    array(
+        'eventname'   => '\core\event\course_restored',
+        'callback'    => 'format_grid_observer::course_restored',
+    )
 );