+++ Changes.md
@@ -5,6 +5,7 @@ Version 3.7.0.2
   2. Add missing Moodle defined.
   3. Styles tidy up.
   4. Removed old convert_legacy_image.php script as was only needed for conversion from M1.9 grid courses - #93.
+  5. Fix cannot change image tiles when importing a grid format course - #66.
 
 Version 3.7.0.1
   1. Beta version for Moodle 3.7 based on 3.6.0.5.